tests/wpt/web-platform-tests/eventsource/format-field-event-empty.htm

<!doctype html>
<html>
  <head>
    <title>EventSource: empty "event" field</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var source = new EventSource("resources/message.py?message=event%3A%20%0Adata%3Adata")
        source.onmessage = function(e) {
          test.step(function() {
            assert_equals("data", e.data)
            this.close()
          }, this)
          test.done()
        }
      })
    </script>
  </body>
</html>
